Bug 904105 - part 1: using argument instead of module
previously we wrote config module by makefile and require it on
each module in GAIA_DIR/build. it would be great if using argument for
build config because we can dynamicly change build config when we use
those modules.

e.g., if using those module in firefox exntesion, user can change build
config in a web form if we use argument, but it's difficult by module.